페이지 메뉴의 최대크기 지정, 페이지 이동 시 게시판정보가 누락되지 않도록

+++ src/main/java/com/khg/exam/demo/controller/UsrArticleController.java
@@ -75,7 +75,7 @@ public class UsrArticleController {
 		// 한페이지당 10개씩
 		// 글 20개면 --> 2페이지
 		// 글 24개면 --> 3페이지
-		int pagesCount = (int) Math.ceil(articlesCount / itemsInAPage);
+		int pagesCount = (int) Math.ceil((double) articlesCount / itemsInAPage);
 
 		List<Article> articles = articleService.getForPrintArticles(rq.getLoginedMemberId(), boardId, page,
 				itemsInAPage);
